Fix several functions and bugs for porting CPython.

+++ scripts/elfconv.sh
@@ -39,8 +39,8 @@ setting() {
   CXX=clang++-16
   CLANGFLAGS="${OPTFLAGS} -std=c++20 -static -I${ROOT_DIR}/backend/remill/include -I${ROOT_DIR}"
   # emscripten
-  EMCC=emcc
-  EMCC_OPTION="-sASYNCIFY=0 -sPTHREAD_POOL_SIZE=0 -pthread -sALLOW_MEMORY_GROWTH -sEXPORT_ES6 -sENVIRONMENT=web,worker $PRELOAD"
+  EMCC=em++
+  EMCC_OPTION="-sASYNCIFY=0 -sINITIAL_MEMORY=536870912 -sSTACK_SIZE=16MB -sPTHREAD_POOL_SIZE=0 -pthread -sALLOW_MEMORY_GROWTH -sEXPORT_ES6 -sENVIRONMENT=web,worker $PRELOAD"
   EMCCFLAGS="${OPTFLAGS} -I${ROOT_DIR}/backend/remill/include -I${ROOT_DIR}"
   # wasi
   WASISDKCC="${WASI_SDK_PATH}/bin/clang++"
@@ -84,11 +84,16 @@ lifting() {
   # copy `elflift` into current directory.
   cp -p ${ROOT_DIR}/build/lifter/elflift ${CUR_DIR}
   
+  dbg_fun_vma=0
+  if [[ -n "${DEBUG_FUNC_ADDR}" ]]; then
+    dbg_fun_vma=${DEBUG_FUNC_ADDR}
+  fi
+
   ${CUR_DIR}/elflift \
   --arch "$2" \
   --bc_out "${CUR_DIR}/${ELFNAME}.bc" \
   --target_elf "${ELFPATH}" \
-  --dbg_fun_cfg "$3" \
+  --dbg_fun_vma "${dbg_fun_vma}" \
   --bitcode_path "$4" \
   --target_arch "${TARGET_ARCH}" \
   --float_exception "${FLOAT_STATUS_FLAG}" \
@@ -102,8 +107,10 @@ lifting() {
   # TEXTIR creates the .ll file
   if [[ -n "${TEXTIR}" ]]; then
     MAINIR="${CUR_DIR}/${ELFNAME}.ll"
-    llvm-dis-${LLVM_VERSION} "${CUR_DIR}/${ELFNAME}.bc" -o "${ELFNAME}.ll"
-    rm "${CUR_DIR}/${ELFNAME}.bc"
+    if [[ -f "${CUR_DIR}/${ELFNAME}.bc" ]]; then
+      llvm-dis-${LLVM_VERSION} "${CUR_DIR}/${ELFNAME}.bc" -o "${ELFNAME}.ll"
+      rm "${CUR_DIR}/${ELFNAME}.bc"
+    fi
     echo -e "[${GREEN}INFO${NC}] built ${ELFNAME}.ll "
   fi
 
@@ -209,12 +216,13 @@ main() {
   case "${TARGET}" in
     *-native)
       echo -e "[${GREEN}INFO${NC}] Compiling to Native binary (for ${HOST_CPU})... "
+      RUNTIME_MACRO="${RUNTIME_MACRO} -DELFNAME=\"${ELFNAME}\""
       MAINOBJ="${ELFNAME}.o"
-      
+
       if [[ -z "${NO_COMPILED}" ]]; then
         ${CXX} ${CLANGFLAGS} ${RUNTIME_MACRO} -c ${MAINIR} -o ${MAINOBJ}
       fi
-      
+
       ${CXX} ${CLANGFLAGS} ${RUNTIME_MACRO} -o "${ELFNAME}.${HOST_CPU}" ${MAINOBJ} ${ELFCONV_COMMON_RUNTIMES} ${RUNTIME_DIR}/syscalls/SyscallNative.cpp
       echo -e " [${GREEN}INFO${NC}] built ${ELFNAME}.${HOST_CPU}"
       if [[ -n "${OUT_EXE}" ]]; then
@@ -223,13 +231,13 @@ main() {
       return 0
     ;;
     *-wasm)
-      RUNTIME_MACRO="${RUNTIME_MACRO} -DTARGET_IS_BROWSER=1"
+      RUNTIME_MACRO="${RUNTIME_MACRO} -DTARGET_IS_BROWSER=1 -DELFNAME=\"${ELFNAME}\""
       MAINOBJ="${CUR_DIR}/${ELFNAME}.wasm.o"
       PRELOAD=
       MAINGENJS="${CUR_DIR}/${ELFNAME}.generated.js"
       
-      if [[ -n "${MOUNT_DIR}" ]]; then
-        PRELOAD="--preload-file ${MOUNT_DIR}"
+      if [[ -n "${MOUNT_SETTING}" ]]; then
+        PRELOAD="--preload-file ${MOUNT_SETTING}"
       fi
       
       if [[ -z "${NO_COMPILED}" ]]; then
@@ -239,8 +247,8 @@ main() {
         echo -e "[${GREEN}INFO${NC}] NO_COPMILED is ON."
       fi
 
-      # compiles wasm
-      ${EMCC} ${EMCCFLAGS} ${RUNTIME_MACRO} ${EMCC_OPTION} -o ${MAINGENJS} ${MAINOBJ} ${ELFCONV_COMMON_RUNTIMES} ${RUNTIME_DIR}/syscalls/SyscallBrowser.cpp
+      # creates wasm
+      ${EMCC} ${EMCCFLAGS} ${RUNTIME_MACRO} ${EMCC_OPTION} ${PRELOAD} -o ${MAINGENJS} ${MAINOBJ} ${ELFCONV_COMMON_RUNTIMES} ${RUNTIME_DIR}/syscalls/SyscallBrowser.cpp
       echo -e "[${GREEN}INFO${NC}] built ${ELFNAME}.wasm and ${ELFNAME}.js and ${ELFNAME}.html."
       
       # prepare Js and Wasm
@@ -249,13 +257,13 @@ main() {
       return 0
     ;;
     *-wasi32)
-      RUNTIME_MACRO="${RUNTIME_MACRO} -DTARGET_IS_WASI=1"
+      RUNTIME_MACRO="${RUNTIME_MACRO} -DTARGET_IS_WASI=1 -DELFNAME=\"${ELFNAME}\""
       MAINOBJ="${ELFNAME}.wasi.o"
 
       if [[ -z "${NO_COMPILED}" ]]; then
         ${WASISDKCC} ${WASISDKFLAGS} ${RUNTIME_MACRO} -c ${MAINIR} -o ${MAINOBJ}
       fi
-      
+
       echo -e "[${GREEN}INFO${NC}] Compiling to Wasm (for WASI)... "
       ${WASISDKCC} ${WASISDKFLAGS} ${WASISDK_LINKFLAGS} ${RUNTIME_MACRO} -o "${ELFNAME}.wasm" ${MAINOBJ} ${ELFCONV_COMMON_RUNTIMES} ${RUNTIME_DIR}/syscalls/SyscallWasi.cpp
       echo -e "[${GREEN}INFO${NC}] built ${ELFNAME}.wasm"
